Made it playable
Made it so you give it a board state and it will tell you the best move for the model to make and give you the updated board state. In case of more than one best move it randomly selects one, and in case of no best move it finds a move that will tie. In case of no tie moves it then randomly selects a move to take.